backend/requirements.txt
fastapi
uvicorn
pydantic
requests
httpx
google-generativeai
motor
python-dotenv